Uso proposed solution also on mobile
