+++ src/components/Ritiro/OperatorsList.tsx
@@ -1,5 +1,3 @@
-import * as React from "react";
-
 import {
   ListItem,
   Typography,
@@ -10,7 +8,7 @@ import {
   TablePagination,
 } from "@mui/material";
 import { RaddOperator } from "../../model";
-import { useRef, useState } from "react";
+import { useEffect, useRef, useState } from "react";
 import CustomPagination from "../CustomPagination";
 
 type Props = {
@@ -22,6 +20,17 @@ function OperatorsList({ rows }: Readonly<Props>) {
   const [rowsPerPage, setRowsPerPage] = useState(10);
 
   const listContainerRef = useRef<HTMLUListElement | null>(null);
+
+  const resizeParentIframe = () => {
+    // This function is to be called if parentIframe.resize() does not work
+    // Current use cases: when changing pagination size, when changing to page with less/more elements
+    if (window.parentIframe) {
+      window.parentIframe.sendMessage({
+        type: "resize",
+        newChildHeight: document.body.scrollHeight,
+      });
+    }
+  };
 
   const handleChangePage = (_event: any, newPage: number | null) => {
     if (newPage !== null) {
@@ -36,12 +45,15 @@ function OperatorsList({ rows }: Readonly<Props>) {
     setPage(0);
     setRowsPerPage(parseInt(event.target.value, 10));
   };
+
   const pagination = {
     size: rowsPerPage,
     totalElements: rows.length,
     numOfDisplayedPages: Math.min(Math.ceil(rows.length / rowsPerPage), 3),
     currentPage: page,
   };
+
+  useEffect(resizeParentIframe, [rowsPerPage, page]);
 
   return (
     <Stack>
@@ -103,7 +115,7 @@ function OperatorsList({ rows }: Readonly<Props>) {
           page={page}
           onPageChange={handleChangePage}
           onRowsPerPageChange={handleChangeRowsPerPage}
-          rowsPerPageOptions={[10]}
+          rowsPerPageOptions={[10, 30, 50]}
         />
         <CustomPagination pagination={pagination} onChange={handleChangePage} />
       </Stack>
