+++ src/index.d.ts
@@ -1,0 +1,12 @@
+export {};
+
+declare global {
+  interface Window {
+    parentIframe?: {
+      sendMessage: (message: {
+        type: "resize";
+        newChildHeight: number;
+      }) => void;
+    };
+  }
+}
